+++ src/App.tsx
@@ -1,30 +1,11 @@
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import './App.scss';
 import { peopleFromServer } from './data/people';
 import { Person } from './types/Person';
-import classNames from 'classnames';
+import { Autocomplete } from './Autocomplete';
 
-type Props = {
-  delay?: number;
-};
-
-export const App: React.FC<Props> = ({ delay = 300 }) => {
-  const [query, setQuery] = useState('');
+export const App: React.FC = () => {
   const [selectedPerson, setSelectedPerson] = useState<Person | null>(null);
-  const [isOpen, setIsOpen] = useState(false);
-  const [appliedQuery, setappliedQuery] = useState('');
-
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      setappliedQuery(query);
-    }, delay);
-
-    return () => clearTimeout(timer);
-  }, [query, delay]);
-
-  const filteredPerson = peopleFromServer.filter(person =>
-    person.name.toLocaleLowerCase().includes(appliedQuery.toLocaleLowerCase()),
-  );
 
   return (
     <div className="container">
@@ -38,57 +19,11 @@ export const App: React.FC<Props> = ({ delay = 300 }) => {
             {`${selectedPerson.name} (${selectedPerson.born} - ${selectedPerson.died})`}
           </h1>
         )}
-        <div className={classNames('dropdown', { 'is-active': isOpen })}>
-          <div className="dropdown-trigger">
-            <input
-              type="text"
-              placeholder="Enter a part of the name"
-              className="input"
-              data-cy="search-input"
-              value={query}
-              onChange={event => {
-                setQuery(event.target.value);
-                setSelectedPerson(null);
-              }}
-              onFocus={() => setIsOpen(true)}
-            />
-          </div>
-
-          <div className="dropdown-menu" role="menu" data-cy="suggestions-list">
-            <div className="dropdown-content">
-              {filteredPerson.map(person => (
-                <div
-                  className="dropdown-item"
-                  data-cy="suggestion-item"
-                  key={person.name}
-                  onClick={() => {
-                    setSelectedPerson(person);
-                    setQuery(person.name);
-                    setIsOpen(false);
-                  }}
-                >
-                  <p className="has-text-link">{person.name}</p>
-                </div>
-              ))}
-            </div>
-          </div>
-        </div>
-
-        {isOpen && filteredPerson.length === 0 && (
-          <div
-            className="
-            notification
-            is-danger
-            is-light
-            mt-3
-            is-align-self-flex-start
-          "
-            role="alert"
-            data-cy="no-suggestions-message"
-          >
-            <p className="has-text-danger">No matching suggestions</p>
-          </div>
-        )}
+        <Autocomplete
+          persons={peopleFromServer}
+          delay={300}
+          onSelected={setSelectedPerson}
+        />
       </main>
     </div>
   );
